and another creator bites the dust
